Add kernel.org's `.clang-format` for editor-agnostic C formatting hints
`clang-format` is an editor-agnostic means of describing C formatting prefs
for code editors.

Since upstream kernel uses this to define an explicit contract for C
formatting, and our BPF code follows the kernel style, this just brings
in a copy of the kernel `.clang-format` dotfile.

This will get picked up by any editor that supports it
(VSCode w/ C extension, emacs with `lsp`, vims, etc).

This will help contributors track the kernel C style for eBPF code,
as well as help this project explicitly declare any deviations from
that style.

This is a straight copy of the
[kernel.org](https://github.com/torvalds/linux/blob/master/.clang-format)
version, with a kernel-specific macro allowlist removed.

Signed-off-by: Benjamin Leggett <benjamin.leggett@solo.io>